Revise functionality and naming for reservoir sampling and created a demo that alters the reservoir.
Changes for reservoir sampling:

Function name change: reservoir_sample -> reservoir_iterator

Reverted the definition of get method to standard version, returning current item of iterator. Returning items from the reservoir with get will require a redesign.

Lower number of data used in demo.

Add code to demo so that a reservoir is filled and then two more items are considered. One of which is added to the reservoir and one of which isn't (based upon the algorithm).