Two tests are added for ReservoirIterable adaptor. Both tests pass.
One test asserts that no replacements are made to the initial reservoir if the probabilities of subsequent items are close to 0.

The other test asserts that the initial reservoir is fully replaced if all of the subsequent items have probabilities close to 1.

The explanation of the wrs_demo is revised.

The extraneous unused test of a deprecated method, get_reservoir, is deleted.