Fix bugs and add tests for ReservoirIterable.
Bug in advance method of ReservoirIterable fixed. Previously the initial call to next() involved filling the reservoir with the first <capacity> elements of the stream and then possibly (since the algorithm is random) replacing one with the next item in the stream. Now the first call to next() fills the reservoir and returns that initial reservoir.

Two tests are added for ReservoirIterable adaptor:

1) One test asserts that no replacements are made to the initial reservoir if the probabilities of subsequent items are close to 0.

2) The other test asserts that the initial reservoir is fully replaced if all of the subsequent items have probabilities close to 1.

fill_reservoir_test(): Now reflects current functionality in which a single call to next fills the reservoir. Removed need for clone.

All tests pass.