Bug fix for advance method of ReservoirIterable.
Previously the case in which capacity > stream_length resulted in an infinite loop. This is fixed.

Test added to show that the case of capacity > stream_length is handled correctly. Test passes.

Switch WeightedDatum{} initialization to using new_datum.

Generation of a stream with weights that result in constant probability: formula for weights simplified. There is only a single formula needed for all weights.