Moments test for WRS now works. Only uses the mean. Threshold is still arbitrary.
Moments test: It is asserted that the moments (currently only the mean) of the full stream and the reservoir are within a threshold.

The test passes sometimes.

Documentation of the fail rate is needed.

The threshold needs to be defined in terms of a confidence interval.

Higher moments need to be added.