Wd wrapper (#21)
Add WDIterable and ExtractValue iterable and their associated constructor, methods, demos, and tests. 

WDIterable wraps the items of an iterator in WeightedDatum. The main use case is to be able to apply Weighted Reservoir Sampling. A demo is provided.

ExtractValue adapts a WDIterable to have items that forget the weight and consist only of the value. 

These are tested by showing that applying ExtractValue after WDIterable to an iterator results in the original iterator.