Reservoir sampling (#25)
Add Reservoir Sampling using ReservoirIterable (no weights) updating old names appropriately.

A reservoir sampling algorithm is added. It uses the struct ReservoirIterable, which is distinct from the previous struct with this name. The advance and get methods are implemented. The previous struct ReservoirIterable is renamed WeightedReservoirIterable and associated functions are updated in lib and examples. All tests pass.

Change names of test functions for WRS to use word weighted.

Add fill reservoir test for ReservoirIterable.

Add replace initial reservoir test for ReservoirIterable. Test passes but needs to be documented better and parameters modified. This test is probabilistic and so will fail occassionally. The parameters should be chosen to achieve a low fail rate and an estimate of the fail rate should be documented.

Rename oracle -> rng for (weighted) reservoir sampling.